+++ packages/terra-action-footer/tests/wdio/theme/block-action-footer-theme-spec.js
@@ -5,13 +5,13 @@ const {
 
 describe('Block action footer theme', () => {
   describe('No Actions', () => {
-    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/without-actions-action-footer-container'));
+    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/without-actions-block-action-footer'));
 
     Terra.should.themeEachCustomProperty(withoutActionsThemeables);
   });
 
   describe('With Actions', () => {
-    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/with-actions-action-footer-container'));
+    beforeEach(() => browser.url('/#/raw/tests/terra-action-footer/action-footer/with-actions-block-action-footer'));
 
     Terra.should.themeEachCustomProperty(withActionsThemeables);
   });