+++ packages/terra-breakpoints/src/terra-dev-site/doc/example/ActiveBreakpointProviderExample.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+
+// eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
+import { ActiveBreakpointProvider, withActiveBreakpoint, ActiveBreakpointContext } from 'terra-breakpoints';
+
+const ActiveBreakpointConsumer1 = withActiveBreakpoint(({ activeBreakpoint }) => (
+  <div>
+    <p>This component uses the withActiveBreakpoint() component generator to interface with the ActiveBreakpointProvider.</p>
+    <p>
+      The active breakpoint is:
+      {' '}
+      {activeBreakpoint}
+    </p>
+  </div>
+));
+
+const ActiveBreakpointConsumer2 = () => (
+  <ActiveBreakpointContext.Consumer>
+    {
+      activeBreakpoint => (
+        <div>
+          <p>This component interfaces with the ActiveBreakpointContext directly.</p>
+          <p>
+            The active breakpoint is:
+            {' '}
+            {activeBreakpoint}
+          </p>
+        </div>
+      )
+    }
+  </ActiveBreakpointContext.Consumer>
+);
+
+const ActiveBreakpointProviderExample = () => (
+  <ActiveBreakpointProvider>
+    <ActiveBreakpointConsumer1 />
+    <hr />
+    <ActiveBreakpointConsumer2 />
+  </ActiveBreakpointProvider>
+);
+
+export default ActiveBreakpointProviderExample;