+++ packages/terra-form-field/CHANGELOG.md
@@ -4,6 +4,14 @@ ChangeLog
 Unreleased
 ----------
 
+2.31.0 - (November 19, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* maxWidth prop added to Field component
+
 2.30.0 - (November 13, 2018)
 ------------------
 ### Changed