+++ packages/terra-form-input/CHANGELOG.md
@@ -4,6 +4,14 @@ ChangeLog
 Unreleased
 ----------
 
+1.31.0 - (November 19, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Dev site test for maxWidth prop
+
 1.30.0 - (November 13, 2018)
 ------------------
 ### Changed