+++ packages/terra-legacy-theme/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.27.0 - (November 19, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.26.0 - (November 13, 2018)
 ------------------
 ### Changed