+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.28.0 - (November 19, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.27.0 - (November 14, 2018)
 ------------------
 ### Changed