+++ packages/terra-props-table/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.22.0 - (November 19, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.21.0 - (November 13, 2018)
 ------------------
 ### Changed